+++ src/selectors/locales.js
@@ -3,7 +3,7 @@ import { get } from 'lodash';
 import locales from '../locales';
 
 /*
- * get content from locales, currently hard-coded to enUS.
+ * get content from locales, currently hard-coded to en-US.
  */
 export const getContent = (state, contentId) =>
   get(locales, [state.locale, contentId], '');
